Add a nonce generator, initially to be used for critical selector beaconing. This just plumbs the nonce generator into the rest of our infrastructure; I decided to split off the actual *use* of the nonce generator in beaconing into a separate revision for your sanity and mine.  This one touches a lot of files fairly minimally; the next will touch a few files but make some largish changes (especially in tests).

+++ src/net/instaweb/rewriter/rewrite_driver.cc
@@ -2667,8 +2667,7 @@ OutputResourcePtr RewriteDriver::CreateOutputResourceWithPath(
       server_context_, mapped_path, unmapped_path, base_url,
       full_name, options(), kind));
 
-  if (options()->max_url_size()
-      < (static_cast<int>(resource->url().size()) + extra_len)) {
+  if (options()->max_url_size() < (resource->url().size() + extra_len)) {
     resource.clear();
     return resource;
   }
